Revert "Move finalize to lib/finalize"
