+++ Cargo.toml
@@ -34,7 +34,6 @@ members = [
     "lib/configurable-derive",
     "lib/condition",
     "lib/event",
-    "lib/finalize",
     "lib/framework",
     "lib/multiline",
     "lib/serde_regex",
@@ -346,7 +345,6 @@ codecs = { path = "lib/codecs" }
 condition = { path = "lib/condition", optional = true }
 configurable = { path = "lib/configurable" }
 event = { path = "lib/event" }
-finalize = { path = "lib/finalize" }
 framework = { path = "lib/framework" }
 humanize = { git = "https://github.com/f1shl3gs/humanize", default-features = false, features = ["bytes", "serde"], rev = "88909181cae00e859a646162cb2521b40006b526" }
 jaeger = { path = "lib/jaeger", optional = true }
