+++ lib/event/src/log/mod.rs
@@ -5,18 +5,18 @@ use std::fmt::Debug;
 
 use bytes::Bytes;
 use chrono::Utc;
-use finalize::{BatchNotifier, EventFinalizer, EventFinalizers, Finalizable};
 use log_schema::log_schema;
 use lookup::Path;
 use measurable::ByteSizeOf;
 use serde::{Deserialize, Serialize};
 use tracing::field::Field;
-use value::keys::{all_fields, keys};
 pub use value::Value;
 
 use crate::metadata::EventMetadata;
 use crate::tags::{skip_serializing_if_empty, Key, Tags};
-use crate::{EventDataEq, MaybeAsLogMut};
+use crate::MaybeAsLogMut;
+use crate::{BatchNotifier, EventDataEq, EventFinalizer, EventFinalizers, Finalizable};
+use value::keys::{all_fields, keys};
 
 /// The type alias for an array of `LogRecord` elements
 pub type Logs = Vec<LogRecord>;
