--- lib/finalize/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "finalize"
-version = "0.1.0"
-edition = "2021"
-publish = false
-
-[dependencies]
-async-stream = { version = "0.3.3" }
-crossbeam-utils = { version = "0.8.14", default-features = false }
-futures = { version = "0.3.26", default-features = false }
-measurable = { path = "../measurable" }
-tokio = { version = "1.25.0", default-features = false }
-pin-project-lite = { version = "0.2.9", default-features = false }
