--- lib/finalize/src/lib.rs
@@ -1,5 +1,0 @@
-mod finalization;
-pub use finalization::*;
-
-mod finalizer;
-pub use finalizer::*;
