+++ lib/framework/src/source/util/mod.rs
@@ -1,8 +1,10 @@
+mod finalizer;
 pub mod http;
 mod tcp;
 mod unix_stream;
 mod wrappers;
 
 // re-export
+pub use finalizer::OrderedFinalizer;
 pub use tcp::{SocketListenAddr, TcpNullAcker, TcpSource};
 pub use unix_stream::build_unix_stream_source;
