+++ lib/testify/Cargo.toml
@@ -7,7 +7,6 @@ publish = false
 [dependencies]
 event = { path = "../event" }
 hyper = { version = "0.14.24" }
-finalize = { path = "../finalize" }
 futures = { version = "0.3.26" }
 futures-util = { version = "0.3.26" }
 rand = "0.8.5"
