+++ lib/testify/src/random.rs
@@ -1,11 +1,9 @@
-use std::collections::HashMap;
-use std::iter;
-
-use event::{Event, Events, LogRecord};
-use finalize::BatchNotifier;
+use event::{BatchNotifier, Event, Events, LogRecord};
 use futures::{stream, Stream, StreamExt};
 use rand::distributions::Alphanumeric;
 use rand::{thread_rng, Rng};
+use std::collections::HashMap;
+use std::iter;
 
 pub fn random_string(len: usize) -> String {
     thread_rng()
