+++ src/sources/kafka.rs
@@ -9,10 +9,10 @@ use codecs::decoding::{DeserializerConfig, FramingConfig, StreamDecodingError};
 use codecs::{Decoder, DecodingConfig};
 use configurable::configurable_component;
 use event::{log::Value, BatchNotifier, BatchStatus, Event, LogRecord};
-use finalize::OrderedFinalizer;
 use framework::config::{DataType, Output, SourceConfig, SourceContext};
 use framework::pipeline::Pipeline;
 use framework::shutdown::ShutdownSignal;
+use framework::source::util::OrderedFinalizer;
 use framework::{Error, Source};
 use futures::{Stream, StreamExt};
 use log_schema::{log_schema, LogSchema};
