+++ src/sources/tail/mod.rs
@@ -10,8 +10,8 @@ use chrono::Utc;
 use configurable::{configurable_component, Configurable};
 use encoding_transcode::{Decoder, Encoder};
 use event::{fields, tags, BatchNotifier, BatchStatus, Event, LogRecord};
-use finalize::OrderedFinalizer;
 use framework::config::{DataType, Output, SourceConfig, SourceContext};
+use framework::source::util::OrderedFinalizer;
 use framework::{hostname, Pipeline, ShutdownSignal, Source};
 use futures::Stream;
 use futures_util::{FutureExt, StreamExt, TryFutureExt};
